requirement.txt
absl-py==2.1.0
albumentations==1.4.0
asttokens==2.4.1
astunparse==1.6.3
blinker==1.9.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.10.0
debugpy==1.6.7
decorator==5.1.1
deep-sort-realtime==1.3.2
defusedxml==0.7.1
dill==0.3.9
easyocr==1.7.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flask==3.1.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.6.0
gitdb==4.0.11
gitpython==3.1.43
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
huggingface-hub==0.26.3
idna==2.10
imageio==2.34.0
importlib-metadata==7.1.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.5.0
keras==3.5.0
kiwisolver==1.4.5
lapx==0.5.11.post1
lazy-loader==0.4
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numpy==1.26.4
onnx==1.16.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.42
protobuf==3.20.3
psutil==5.9.0
pure_eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pygments==2.17.2
pyparsing==3.1.2
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pywin32==305
pyyaml==6.0.1
pyzmq==25.1.2
qudida==0.0.4
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.27
safetensors==0.4.5
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==68.2.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
stack_data==0.6.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tf2onnx==1.16.1
thop==0.1.1-2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.20.3
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.2
transformers==4.46.3
typing_extensions==4.9.0
ultralytics==8.3.39
ultralytics-thop==2.0.12
urllib3==2.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
werkzeug==3.1.3
wheel==0.41.2
wrapt==1.16.0
zipp==3.17.0